[@types/react] Improve ComponentProps handling of class components
Wrapping P in ReactManagedAttributes generates a type where
defaultProps of a class component are taken into consideration.
