+++ types/react/tsconfig.json
@@ -4,6 +4,7 @@
         "test/index.ts",
         "test/tsx.tsx",
         "test/cssProperties.tsx",
+        "test/ComponentProps.tsx",
         "test/managedAttributes.tsx",
         "test/hooks.tsx"
     ],
